LU: Fill in compute, apply
Fill in the "main" part of task 3; the LU class.  The `compute' function
computes the LU decomposition of a matrix with the most straightforward
algorithm possible (no pivot) and the `apply' function solves a linear
system A · x = b for x.  It does this by forwards and backwards
application of LU factors.