Add crs-matrix.hh
This is the declarative header file for exercise 5, task 4.

For the compression, I went with returning an different class instead of
changing the base object.  This has the advantage that one does not need
even more state to check whether the given matrix is compressed.  Thus,
operations like modify (and variations) don't need to check _even more_
preconditions.

Invariably, since we can't destroy the original object, this leads to us
keeping the same matrix in memory twice.  This may or may not be that
bad, as one would expect the compressed vector to be much smaller than
even the sparse vector (as it is not sprinkled with zeros).

Plus, there is also a silver lining: one can now compress the matrix and
then still work with it afterwards—this would normally require a more
complex implementation of modify.